+++ test-app/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @floatdesk/test-app
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @floatdesk/react@2.2.0
+  - @floatdesk/sdk@2.2.0
+
 ## 0.0.2
 
 ### Patch Changes